WIP fixes

+++ server/encoder/video_encoder_vulkan.h
@@ -87,7 +87,7 @@ class video_encoder_vulkan : public xrt::drivers::wivrn::VideoEncoder
 	vk::VideoEncodeRateControlLayerInfoKHR rate_control_layer;
 
 protected:
-	const uint32_t num_dpb_slots = 2;
+	const uint8_t num_dpb_slots = 2;
 	std::optional<vk::VideoEncodeRateControlInfoKHR> rate_control;
 
 	video_encoder_vulkan(wivrn_vk_bundle & vk, vk::Rect2D rect, vk::VideoEncodeCapabilitiesKHR encode_caps, float fps, uint64_t bitrate);
